Type fixes

+++ HISSTools_IR_Toolbox_Dependencies/HIRT_Multichannel_Convolution/Convolver.cpp
@@ -127,7 +127,7 @@ ConvolveError HISSTools::Convolver::set(uint32_t inChan, uint32_t outChan, const
     std::vector<float> inputFloat(impulseLength);
     
     for (unsigned long i = 0; i < impulseLength; i++)
-        inputFloat[i] = input[i];
+        inputFloat[i] = static_cast<float>(input[i]);
     
     return set(inChan, outChan, inputFloat.data(), impulseLength, resize);
 }
